chat_with_file/requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
google-generativeai>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
docx2txt>=0.8
unstructured>=0.13.0
python-dotenv>=1.0.0
pandas>=2.0.0
tiktoken>=0.7.0
